refactor: 각 기물에 맞는 전략을 가지도록 수정
