+++ src/main/java/chess/domain/piece/King.java
@@ -1,23 +1,11 @@
 package chess.domain.piece;
 
-import chess.domain.board.Square;
-import java.util.ArrayList;
-import java.util.List;
+import chess.domain.piece.strategy.KingStrategy;
 
 public class King extends Piece {
 
     public King(final Color color) {
-        super(color);
-    }
-
-    @Override
-    public List<Square> findRoute(final Square source, final Square destination) {
-        final int distanceFile = destination.calculateDistanceFile(source);
-        final int distanceRank = destination.calculateDistanceRank(source);
-        if (!KingVector.isExistMovableVector(distanceFile, distanceRank)) {
-            throw new IllegalArgumentException("해당 기물이 움직일 수 있는 경로가 아닙니다.");
-        }
-        return new ArrayList<>(List.of(destination));
+        super(color, new KingStrategy());
     }
 
     @Override
