+++ src/main/java/chess/domain/piece/strategy/Strategy.java
@@ -1,0 +1,9 @@
+package chess.domain.piece.strategy;
+
+import chess.domain.board.Square;
+import java.util.List;
+
+public interface Strategy {
+
+    List<Square> findRoute(Square source, Square destination);
+}
