+++ src/test/java/chess/util/TestStrategy.java
@@ -1,0 +1,13 @@
+package chess.util;
+
+import chess.domain.board.Square;
+import chess.domain.piece.strategy.Strategy;
+import java.util.List;
+
+public class TestStrategy implements Strategy {
+
+    @Override
+    public List<Square> findRoute(final Square source, final Square destination) {
+        return null;
+    }
+}
